Changed project's struct
