+++ Navigation.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:">
+      location = "self:/Users/kkrsvn/Documents/ios-homeworks/Navigation/Navigation.xcodeproj">
    </FileRef>
 </Workspace>
