--- Navigation/Post.swift
@@ -1,12 +1,0 @@
-//
-//  Post.swift
-//  Navigation
-//
-//  Created by Kirill Krasavin on 12.02.2022.
-//
-
-import UIKit
-
-struct Post {
-    let title: String
-}
