+++ Navigation/Presentation/Extention/UIImageAlphaChange.swift
@@ -1,0 +1,18 @@
+//
+//  UIImageAlphaChange.swift
+//  Navigation
+//
+//  Created by Kirill Krasavin on 23.02.2022.
+//
+
+import UIKit
+
+extension UIImage {
+    func image(alpha: CGFloat) -> UIImage? {
+        UIGraphicsBeginImageContextWithOptions(size, false, scale)
+        draw(at: .zero, blendMode: .normal, alpha: alpha)
+        let newImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return newImage
+    }
+}
