+++ Navigation/Presentation/Protocols/LoginViewHeaderDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  LoginViewHeaderDelegate.swift
+//  Navigation
+//
+//  Created by Kirill Krasavin on 23.02.2022.
+//
+
+import Foundation
+
+protocol LoginViewHeaderDelegate: AnyObject {
+    func didTapButton()
+}
