+++ Navigation/Resourses/Assets.xcassets/AppIcon.appiconset/Contents.json
