+++ Navigation/Resourses/Assets.xcassets/blue_pixel.imageset/blue_pixel.png
