+++ Navigation/Resourses/Assets.xcassets/logo.imageset/Contents.json
