+++ Navigation/Resourses/Assets.xcassets/logo.imageset/logo.png
